requirements.txt
discord.py == 1.7.3
gspread == 0.4.1
oauth2client == 2.2.0
numpy == 1.20.1
requests == 2.27.1
urllib3 == 1.26.8